+++ src/config/metainfo.js
@@ -1,5 +1,5 @@
 export default {
-  title: "Id More Academy - Tech from Engineer Perspective by Id More Team",
+  title: "Yussan Academy - Tech from Engineer Perspective by Id More Team",
   description:
     "Here we are not only focused on making tech products. But it also makes technology accessible, affordable and easy for everyone to learn."
 }